==> Natural Language Processing/notebooks/2-Document-Term Matrix.ipynb <==
# IT Tickets Classification Project

## Previous Notebooks

- [Data Collection](0-Data Collection.ipynb)
- [Data Cleaning and EDA](1-Data Cleaning and EDA.ipynb)

import numpy as np
import pandas as pd
from sklearn.feature_extraction.text import TfidfTransformer
from scipy.sparse.linalg import svds

## Document-Term Matrix

In this notebook I will transform the tickets' bag of words into a document-term matrix, which I'll later use to do some topic modeling.

In doing so I will set aside all the words that appear in every ticket or just once.

bag = pd.read_pickle('../data/interim/bag_of_words.pkl')

n = len(bag)
n

all_terms = {}
for row in bag['bag_of_words']:
    for term in row:
        if term in all_terms:
            all_terms[term] += row[term]
        else:
            all_terms[term] = row[term]

m = len(all_terms)
m

terms = {}
for term in all_terms:
    if all_terms[term] > 1 and all_terms[term] < m:
        terms[term] = all_terms[term]

m = len(terms)
m

doc_term = pd.DataFrame(np.zeros([n, m]), columns=list(terms), index=bag['key'])
doc_term['bag_of_words'] = bag[['bag_of_words', 'key']].set_index('key')

for col in doc_term.columns:
    doc_term[col] = [bag[col] if col in bag else 0 for bag in doc_term['bag_of_words']]

doc_term.drop('bag_of_words', axis=1, inplace=True)

for col in doc_term.columns:
    if len(doc_term.loc[doc_term[col]!=0]) == 1 or len(doc_term.loc[doc_term[col]!=0]) == n:
        doc_term.drop(col, axis=1, inplace=True)

doc_term.head()

doc_term.to_pickle('../data/interim/document_term.pkl')

## Following Notebooks

- [Topic Modeling](3-Topic Modeling.ipynb)
- [Random Forest Prediction](4-Model.ipynb)
