==> Natural Language Processing/notebooks/3-Topic Modeling.ipynb <==
# IT Tickets Classification Project

## Previous Notebooks

- [Data Collection](0-Data Collection.ipynb)
- [Data Cleaning and EDA](1-Data Cleaning and EDA.ipynb)
- [Document-Term Matrix](2-Document-Term Matrix.ipynb)

import numpy as np
import pandas as pd
from sklearn.feature_extraction.text import TfidfTransformer
from scipy.sparse.linalg import svds

## Topic Modeling

In this notebook I use SVD to do topic modeling of the tickets' text, selecting the first 20 eigenvectors to represent the 20 most relevant topics in the collection. Before applying SVD I normalize the document-term matrix using TF-IDF.

doc_term = pd.read_pickle('../data/interim/document_term.pkl')

tfidf = TfidfTransformer()
X = tfidf.fit_transform(doc_term)

k = 20
u, s, vt = svds(X, k=k)

s

Below are the 10 most relevant keywords of the selected topics: topic 20 seems to be related to some technical terms, while the other topics seem to be related to agreements (e.g. topic 16, 15 and 8), suppliers (e.g. topic 19, 18 and 11) or IT procedures (e.g. topic 12 and 11).

topics = []
for i in range(k):
    topics.append(list(pd.Series(vt[i], index=doc_term.columns).sort_values(ascending=False).head(10).index.values))

topics = pd.DataFrame(topics)
topics.columns = ['word{}'.format(i+1) for i in topics.columns.values]
topics.sort_index(ascending=False, inplace=True)
topics.index = ['topic{}'.format(i+1) for i in topics.index.values]
topics.T.iloc[:, 0:10]

topics.T.iloc[:, 10:20]

issues = pd.read_pickle('../data/interim/bag_of_words.pkl')

Finally I'm going to create the dataset I will use to predict the topics' labels by adding to the topic data the label and issue type:

processed_data = pd.DataFrame(u, index=issues['key'])
processed_data.columns = ['topic{}'.format(i+1) for i in processed_data.columns.values]
processed_data = processed_data.merge(issues[['key', 'issue_type', 'label']].set_index('key'),
                                      how='inner', left_index=True, right_index=True)

processed_data.head()

processed_data.to_pickle('../data/processed/proc_data.pkl')

## Following Notebooks

- [Random Forest Prediction](4-Model.ipynb)
