==> Natural Language Processing/notebooks/4-Model.ipynb <==
# IT Tickets Classification Project

## Previous Notebooks

- [Data Collection](0-Data Collection.ipynb)
- [Data Cleaning and EDA](1-Data Cleaning and EDA.ipynb)
- [Document-Term Matrix](2-Document-Term Matrix.ipynb)
- [Topic Modeling](3-Topic Modeling.ipynb)

import numpy as np
import pandas as pd
from sklearn.preprocessing import LabelEncoder
from sklearn.ensemble import RandomForestClassifier
from sklearn.model_selection import train_test_split, GridSearchCV, StratifiedKFold
from sklearn.metrics import f1_score, precision_recall_curve, confusion_matrix, accuracy_score, recall_score, precision_score, make_scorer

## Random Forest

In this last notebook I use the topics determined in the previous steps and the issue type to predict the tickets' labels using a random forest.

I try using different numbers of topics and determine the best one using 5-fold cross-validation.

processed_data = pd.read_pickle('../data/processed/proc_data.pkl')

le1 = LabelEncoder()
processed_data['issue_type'] = le1.fit_transform(processed_data['issue_type'])
le2 = LabelEncoder()
processed_data['label'] = le2.fit_transform(processed_data['label'])

Xtr, Xts, ytr, yts = train_test_split(processed_data.drop('label', axis=1),
                                      processed_data['label'],
                                      test_size=0.2, random_state=42, stratify=processed_data['label'])

rfc = RandomForestClassifier()

param_grid = {'n_estimators':np.arange(30, 110, 10),
              'max_features':np.arange(0.1, 1.1, 0.1),
              'min_samples_split': np.arange(2, 6, 1)}
f1scorer = make_scorer(f1_score, average='weighted')
kf = StratifiedKFold(n_splits=5, random_state=74)

gs10 = GridSearchCV(estimator=rfc, param_grid=param_grid, n_jobs=-1, cv=kf, verbose=10, scoring=f1scorer)
gs10.fit(Xtr[['issue_type', 'topic10','topic11', 'topic12', 'topic13', 'topic14',
              'topic15', 'topic16', 'topic17', 'topic18', 'topic19', 'topic20']], ytr)

print(gs10.best_score_)
print(gs10.best_params_)

gs15 = GridSearchCV(estimator=rfc, param_grid=param_grid, n_jobs=-1, cv=kf, verbose=10, scoring=f1scorer)
gs15.fit(Xtr[['issue_type', 'topic5','topic6', 'topic7', 'topic8', 'topic9',
              'topic10','topic11', 'topic12', 'topic13', 'topic14',
              'topic15', 'topic16', 'topic17', 'topic18', 'topic19', 'topic20']], ytr)

print(gs15.best_score_)
print(gs15.best_params_)

gs20 = GridSearchCV(estimator=rfc, param_grid=param_grid, n_jobs=-1, cv=kf, verbose=10, scoring=f1scorer)
gs20.fit(Xtr[['issue_type', 'topic1','topic2', 'topic3', 'topic4', 'topic5',
              'topic5','topic6', 'topic7', 'topic8', 'topic9',
              'topic10','topic11', 'topic12', 'topic13', 'topic14',
              'topic15', 'topic16', 'topic17', 'topic18', 'topic19', 'topic20']], ytr)

print(gs20.best_score_)
print(gs20.best_params_)

The best model is the one with all the 20 most relevant topics in it, even if its cv score is not much greater than the model with 15 topics.

rfc.set_params(max_features=0.3, min_samples_split=2, n_estimators=90)
rfc.fit(Xtr, ytr)

for imp, col in sorted(zip(rfc.feature_importances_, Xtr.columns), reverse=True):
    print('{} importance: {:.5f}'.format(col, imp))

ypred_ts = rfc.predict(Xts)
ypred_ts_proba = rfc.predict_proba(Xts)
print('Accuracy: {}'.format(accuracy_score(yts, ypred_ts)))
print('Recall: {}'.format(recall_score(yts, ypred_ts, average='weighted')))
print('Precision: {}'.format(precision_score(yts, ypred_ts, average='weighted')))
print('F1: {}'.format(f1_score(yts, ypred_ts, average='weighted')))

This model gives me an F1 score of almost 0.7, with very balanced accuracy, precision and recall.

The confusion matrix tells me that the model is not very precise in determining "ISAHD Altro" label, which is quite generic and in general it's confused with many other issues' categories. There are also problems in identifying the least numerous categories, such as "OTHER", "ISAHD Rilascio" and "ISAHD Estrazione effettuata".

pd.DataFrame(confusion_matrix(le2.inverse_transform(yts), le2.inverse_transform(ypred_ts)),
             index=le2.classes_,
             columns=le2.classes_)

## Conclusions

In this project I used some NLP techniques to analyze textual data about incoming IT requests, using topic modeling to try and predict how an issues has to be addressed in order to correctly solve it. Since there are a lot of ways to solve an issue this is a multi-class prediction, in which some of the classes are very difficult to predict correctly since they occur rarely.

The final F1 score is near 0.7, which leaves room for improvement for sure: I could be more precise in eliminating noise words and try to collect more issues in order to have more data for the least frequent categories.

## Further Analysis

If I had more time I would have tried:

- to add more topics to my dataset
- to determine a list of words inherent to the language used in requesting something (such as greetings, thanking and so on)
- to see how the predictions change by adding the name of the person who created the ticket
